redo zorkoid movement

+++ 6_escape_from_zorkon/Point.h
@@ -7,26 +7,26 @@ const double Pi = 3.14159265358979;
 
 class Point {
 public:
-  double x; // rectangular coordinates
-  double y;
-  double r; /// polar coordinates
-  double theta;
-  Point(double x, double y, char s);      // constructor
+  long double x; // rectangular coordinates
+  long double y;
+  long double r; /// polar coordinates
+  long double theta;
+  Point(long double x, long double y, char s);      // constructor
   Point operator+(const Point &); // overload addition operator
   Point operator-(const Point &); // overload subtraction operator
-  Point operator*(const double); // overload multiplication operator where right
+  Point operator*(const long double); // overload multiplication operator where right
                                  // operand is double
-  Point operator/(const double); // overload multiplication operator where right
+  Point operator/(const long double); // overload multiplication operator where right
                                  // operand is double
   Point operator+=(const Point &);
   bool operator==(const Point &);
 
 
-  void setTheta(double);
-  double thetaFromXY();
+  void setTheta(long double);
+  long double thetaFromXY();
   void setPolar();
   void setRectangular();
-  double distFrom(const Point&);
+  long double distFrom(const Point&);
 };
 
 std::ostream &operator<<(std::ostream &, const Point &p);
